inventory_app/requirements.txt
matplotlib==3.8.0
seaborn==0.12.2
pandas==2.0.3
numpy==1.24.3
pillow==10.0.0
opencv-python==4.8.0.76
pyzbar==0.1.9
openpyxl==3.1.2
reportlab==4.0.4
tkcalendar==1.6.1
ttkthemes==3.2.0
pyinstaller==6.3.0
